Create file for key result translation
-Add a new file named key_strike

-Delete timer_control file

-Add get_keyboard_to_gamepad() and count_point() to key_strike file
